Make a custom hook to obtain subcollection data.
